+++ pkgs/firehose/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Remove the `version` pubspec checks (these largely duplicate the feedback
   provided by publishing automation).
+- Run health workflow on all packages if it is changed.
+- Specify Flutter packages in the repo, to only have a single workflow file.
 
 ## 0.9.3